refactor: renaming files and titles for Vila dos Cabanos
